refactor: update paths for GET lambda finctions
