+++ product-service/lambda-functions/getProductById.ts
@@ -4,8 +4,8 @@ import {
   APIGatewayProxyResult,
 } from "aws-lambda";
 import * as AWS from "aws-sdk";
-import { IProduct } from "../types/product.interface";
-import { products } from "../products";
+import { IProduct } from "./product.interface";
+import { products } from "./products";
 import { DocumentClient } from "aws-sdk/clients/dynamodb";
 import {
   NotFoundError,
